requirements.txt
Flask
flask-cors
pytesseract
pillow